requirements.txt
Django>=5.1.7
Pillow>=10.2.0  # For image handling
python-dotenv>=1.0.0  # For environment variables 